requirements.txt
autopep8==1.4
bcrypt==3.1.4
blinker==1.4
Brotli==1.0.7
cachetools==3.0.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
click==6.7
flake8==3.5.0
flake8-docstrings==1.3.0
flake8-polyfill==1.0.2
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
flask-mongoengine==0.9.5
Flask-WTF==0.14.2
fonttools==3.32.0
google-api-core==1.5.2
google-auth==1.6.1
google-cloud-core==0.28.1
google-cloud-translate==1.3.1
googleapis-common-protos==1.5.5
gspread==3.0.1
httplib2==0.12.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mccabe==0.6.1
mongoengine==0.15.3
oauth2client==4.1.3
Pillow==5.3.0
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.3.1
pycparser==2.19
pydocstyle==2.1.1
pyflakes==1.6.0
pymongo==3.7.1
pytz==2018.5
requests==2.20.1
rsa==4.0
six==1.11.0
snowballstemmer==1.2.1
urllib3==1.24.1
Werkzeug==0.14.1
WTForms==2.2.1
